+++ src/payment/strategies/cybersource/cardinal.ts
@@ -1,3 +1,7 @@
+import Address from '../../../address/address';
+import BillingAddress from '../../../billing/billing-address';
+import { CreditCardInstrument } from '../../payment';
+
 export const CardinalSignatureValidationErrors = [100004, 1010, 1011, 1020];
 
 export interface CardinalSDK {
@@ -38,6 +42,15 @@ export interface CardinalSetupCompletedData {
 export interface CardinalModuleState {
     loaded: boolean;
     module: string;
+}
+
+export interface CardinalOrderData {
+    billingAddress: BillingAddress;
+    shippingAddress?: Address;
+    currencyCode: string;
+    id: string;
+    amount: number;
+    paymentData: CreditCardInstrument;
 }
 
 export enum CardinalInitializationType {
@@ -93,7 +106,7 @@ export interface CardinalPartialOrder {
 }
 
 export interface CardinalConsumer {
-    Email1: string;
+    Email1?: string;
     Email2?: string;
     ShippingAddress?: CardinalAddress;
     BillingAddress: CardinalAddress;
