+++ src/payment/strategies/cybersource/index.ts
@@ -2,3 +2,4 @@ export * from './cardinal';
 
 export { default as CyberSourcePaymentStrategy } from './cybersource-payment-strategy';
 export { default as CardinalScriptLoader } from './cardinal-script-loader';
+export { default as CardinalClient } from './cardinal-client';
